refactor AudioRecorder component
